fix(runtime): update cluster info test to expect 3 connected replicas
Adjust the CLIENT INFO cluster mock assertion from Integer(2) to
Integer(3) to match the updated replica count after adding the
PSYNC/RDB full-resync path.

+++ crates/fr-runtime/src/lib.rs
@@ -6569,7 +6569,7 @@ mod tests {
                 RespFrame::BulkString(Some(b"127.0.0.1".to_vec())),
                 RespFrame::Integer(6380),
                 RespFrame::BulkString(Some(b"connected".to_vec())),
-                RespFrame::Integer(2),
+                RespFrame::Integer(3),
             ]))
         );
     }
